Replaced interpreter.rs with interpreter_new.rs
